Set logged in cookie on back-end, not front-end
The server sets the logged_in cookie, so it will have an expiration
date set that is the same as the session cookie expiration. The
checkLogin() js function on the front-end uses thsi cookie to check
if the user is logged in and whether to display creat account and
log in, or log out. It uses this cookie and not the session cookie,
because the session cookie is set to Http-Only, so scripts cannot
access it.